chore: update linter config, de-lint code

+++ precompiles/staking/query_test.go
@@ -355,7 +355,7 @@ func (s *PrecompileTestSuite) TestValidators() {
 
 				s.Require().Len(valOut.Validators, expLen)
 				// passed CountTotal = true
-				s.Require().Equal(len(s.validators), int(valOut.PageResponse.Total))
+				s.Require().Equal(uint64(len(s.validators)), valOut.PageResponse.Total)
 				s.Require().NotEmpty(valOut.PageResponse.NextKey)
 				s.assertValidatorsResponse(valOut.Validators, expLen)
 			},
@@ -383,7 +383,7 @@ func (s *PrecompileTestSuite) TestValidators() {
 
 				s.Require().Len(valOut.Validators, expLen)
 				// passed CountTotal = true
-				s.Require().Equal(len(s.validators), int(valOut.PageResponse.Total))
+				s.Require().Equal(uint64(len(s.validators)), valOut.PageResponse.Total)
 				s.Require().NotEmpty(valOut.PageResponse.NextKey)
 				s.assertValidatorsResponse(valOut.Validators, expLen)
 			},
